Solve 3289

+++ 2125/rust/src/lib.rs
@@ -1,12 +1,7 @@
 impl Solution {
     pub fn number_of_beams(bank: Vec<String>) -> i32 {
-        bank.into_iter()
-            .map(|row| row.chars().filter(|c| *c == '1').count() as i32)
-            .filter(|devices| *devices != 0)
-            .collect::<Vec<_>>()
-            .windows(2)
-            .map(|w| w[0] * w[1])
-            .sum()
+        bank.into_iter().map(|row| row.chars().filter(|c| *c == '1').count()).filter(|devices| *devices != 0)
+            .collect::<Vec<_>>().windows(2).map(|w| w[0] * w[1]).sum()
     }
 }
